requirements.txt
annotated-types==0.7.0
anyio==4.8.0
authlib==1.3.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cohere==5.13.12
cryptography==44.0.1
exceptiongroup==1.2.2
fastapi==0.115.8
fastavro==1.10.0
filelock==3.17.0
fsspec==2025.2.0
grpcio==1.70.0
grpcio-health-checking==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
idna==3.10
lxml==5.3.1
packaging==24.2
protobuf==5.29.3
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pymupdf==1.25.3
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
setuptools==75.8.0
sniffio==1.3.1
starlette==0.45.3
tiktoken==0.9.0
tokenizers==0.21.0
tqdm==4.67.1
types-requests==2.32.0.20241016
typing-extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
validators==0.34.0
watchdog==6.0.0
weaviate-client==4.11.0
